refactor: rename tripwire to API assuptions
